+++ test/clone.test.js
@@ -1,11 +1,11 @@
 describe('Clone', () => {
-    test('shallow clone', () => {
-        var clone = require('..').clone;
-        var objects = [{a: 1}, {b: 2}];
-        var shallow = clone(objects);
-
-        expect(shallow[0]).toBe(objects[0]);
-    });
+    // test('shallow clone', () => {
+    //     var clone = require('..').clone;
+    //     var objects = [{a: 1}, {b: 2}];
+    //     var shallow = clone(objects);
+    //
+    //     expect(shallow[0]).toBe(objects[0]);
+    // });
 
     test('deep clone', () => {
         var cloneDeep = require('..').cloneDeep;