+++ benchmark/suites/camelCase-lodash.bench.js
@@ -1,0 +1,18 @@
+const {add, cycle, suite, complete} = require('benny');
+const _ = require('lodash/fp');
+const camelCase = require('../../').camelCase;
+
+module.exports = suite(
+    'camelCase-Lodash',
+
+    add('Old', () => {
+        _.camelCase('foo_bar');
+    }),
+
+    add('Node-Utils', () => {
+        camelCase('foo_bar');
+    }),
+
+    cycle(),
+    complete()
+);