+++ index.js
@@ -1,14 +1,22 @@
 var nodeUtils = function() {
     var _clone = require('./lib/clone');
 
+    this.merge = _clone.merge;
     this.cloneDeepWith = _clone.cloneDeepWith;
     this.cloneDeep = _clone.cloneDeep;
     this.clone = _clone.clone;
 
-    this.merge = _clone.merge;
-
     this.get = require('./lib/get');
+    this.set = require('./lib/set');
+    this.toPath = require('./lib/toPath');
     this.isEmpty = require('./lib/isEmpty');
+    this.isObject = require('./lib/isObject');
+    this.mapKeys = require('./lib/mapKeys');
+    this.forEach = require('./lib/forEach');
+    this.snakeCase = require('./lib/snakeCase');
+    this.camelCase = require('./lib/camelCase');
+    this.kebabCase = require('./lib/kebabCase');
+    this.debounce = require('./lib/debounce');
 
     this.PGTypes = require('./lib/postgres/PGTypes');
     this.PGConnecter = require('./lib/postgres/PGConnecter');
@@ -19,7 +27,7 @@ var nodeUtils = function() {
 
     this.request = require('./lib/request');
 
-    this.keysToCamel = require('./lib/camelCase');
+    this.keysToCamel = require('./lib/keysToCamel');
 
     this.Timer = require('./lib/timer');