+++ test/postgres.test.js
@@ -11,7 +11,7 @@ describe('Postgres', () => {
     try {
         pgOptions.crypto = require('@abeai/node-crypto').utils.pgCrypto;
     } catch (_) {
-      console.log(_)
+        console.log(_);
     }
 
     var pg = new PGConnecter(pgOptions);
@@ -346,9 +346,12 @@ describe('Postgres', () => {
         });
         describe('Delete', () => {
             test('Delete user1 [.delete()]', async ()=>{
+                var id = user1.id;
 
                 await user1.delete();
+                var deletedUser = await UserAD.findById(id);
 
+                expect(deletedUser).toBe(null);
                 expect(user1.username).toBe(null);
             });
 
@@ -419,11 +422,11 @@ describe('Postgres', () => {
                 });
 
                 test('Create user 4 with 5 properties static with encrypt field phone, hashed password and and set encrypted_profile with encryption off [create()]', async ()=>{
-                  try {
-                      pgOptions.crypto = null;
-                      pg = new PGConnecter(pgOptions);
-                  } catch (_) {
-                  }
+                    try {
+                        pgOptions.crypto = null;
+                        pg = new PGConnecter(pgOptions);
+                    } catch (_) {
+                    }
                     user4 = await UserAD.create(
                         {
                             username: 'user4',
@@ -588,9 +591,12 @@ describe('Postgres', () => {
 
             describe('Delete', () => {
                 test('Delete user1 [.delete()]', async ()=>{
+                    var id = user1.id;
 
                     await user1.delete();
+                    var deletedUser = await UserAD.findById(id);
 
+                    expect(deletedUser).toBe(null);
                     expect(user1.username).toBe(null);
                 });
 
@@ -601,6 +607,15 @@ describe('Postgres', () => {
                         });
 
                     expect(deletedUsers[0].email).toBe('motherlover@lel.com');
+                });
+
+                test('Delete user2 [.deleteById()]', async ()=>{
+                    var id = user2.id;
+
+                    await UserAD.deleteById(id);
+                    var deletedUser = await UserAD.findById(id);
+
+                    expect(deletedUser).toBe(null);
                 });
 
                 test('Delete All static [deleteAll()]', async ()=>{